fix: legacy interrupt working
- It seems to be there is a mistake in the documentation of
the FC1121 card. In chapter 3.2.1 Interrupt it is said that
the global interrupt enable/status is bit 7 but instead it
is bit 0 !!!

- The FC1121 card is also configured to map DC Sycn0/1 interrupts
to PDI interrupt for only 100 ns. With this configuration DC interrupts
cannot be used when using standard OS like Linux or Windows. For processing
inside an IRQ handler this is usually too short to determine the INT source.