ci: reenable basic ccat tests on hardware
With todays test infrastructure the easiest way to test ccat modules is
to netboot an upstream Debian kernel on a real device, compile our out
of tree driver on the device and install it.
If everything works, we should now see a bunch of /dev/ccat* devices.

Now, we can run some of our old tests:
- ./unittest/test-gpio.sh
- ./unittest/test-systemtime.sh
- ./unittest/test-update.sh --dry-run

NOTE: Our old driver doesn't work on all modern devices. So we start
with the working families CX20x0 and CX51x0. CX50x0 is dropped due to
its 32-bit legacy.

Signed-off-by: Filip Jöde <F.Joede@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>