ci: Trigger downstream test-tcos pipeline
Instead of building the rootfs here (and having to adjust the kernel
version manually in the process), we trigger a downstream test-tcos
pipeline which handles that for us.
We now build a test pack that runs inside a test-runner container and
builds and tests the CCAT driver on the device.
This saves us the effort of building the rootfs and setting various
configuration options (i.e noninteractive sudo on the test-device) at
the expense of a longer pipeline duration.
Unlike before, TwinCAT is automatically installed on the test device by
the test-tcos pipeline. Unfortunately, this breaks our test as it
assigns the vfio-pci driver before our ccat driver is even built, so we
stop the system service and unload the vfio-pci driver before the test.

Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>