systemtime: enable the use of multiple systemclocks
When using e.g. more than one CCAT device (like multiple FC1121) each
of them provides a CCAT systemclock. This appends a number to the systemclock
name.

For now we simply use a monotonic counter, which is increased on every
probe and never decreased. This should be good enough for all general
usecases.

ATTENTION: This changes the clocksource name so possible users have to
change from 'ccat' to ccat0' like we do in our test.

Co-authored-by: Robert Burger <robert.burger@dlr.de>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>