ci: use more efficient trigger
The "test-tcos" repository introduced a mechanism[1] to speed up testing by
using prebuild images. We can use this to improve our effectiveness by a
large scale. We trigger 7 pipelines in parallel! So we will save many
build jobs with this.

NOTE: Now, we will run in a netboot environment without /lib/modules so
we have to create that directory. manually.

[1] https://git.beckhoff.dev/beckhoff/test-tcos/-/commit/24f2a0bb2cb7a9a2614187a6c7d08d6eda07ceba

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>