netdev: Stop requesting ISA DMA channels
Well, the times of ISA DMA should be long gone and since IOMMUs, VT-d
and AMD-VI appeared, our standalone ccat driver became problematic.

Previously, we allocated a single memory range and split it manually for
the rx and tx DMA queues in the CCAT. With IOMMUS, VT-d and AMD-VI it is
more reliable to allocate two separate PCI mappings for the rx and tx
queue. On top of it we always assumed the dma_handle and virtual address
returned from dma_alloc_coherent() are aligned in the same way. Well,
thats pretty optimistic but it seems smarter to align the physical
address and take the same offset to align the virtual address
accordingly.

While we are at it we drop the totally useless dma_request() calls and
drop the unused "channel" member from "struct ccat_dma_mem".
Additionally, we drop the unused CCAT_ALIGN_CHANNEL() channel macro,
since we can just use the generic PTR_ALIGN().

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>