tests/ccat-module: unload vfio-pci only if loaded
When running in CI with BHF_CI_LOOP=y without reboots, vfio-pci might
not yet be reloaded. In that case rmmod would fail. To distingiush
between failure and not loaded we check with lsmod first.

NOTE: If this turns out to be a race condition just fall back to " || true"

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>