ci: wait for ccat drivers to be loaded before test
To improve the robustness of our test in automatic CI pipelines, we
introduce a new waiting step after the ccat driver was loaded.
Previously, we would just start testing, making the entire pipeline
pretty dependent on fast driver loading. Well, to be sure and to have
better diagnostics in case something fails, we add a new script, which
just waits for the three most common ccat functions: clocksource, gpio,
netddev and update.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>