tests/ccat-module: nuke all pci driver_overrides
If we want to run our test in a loop (BHF_CI_LOOP=y). We have to
reenable the TcSystemService afterwards. The problem is it will add a
driver_override to our CCAT device(even if it was missing in the first
run). To get rid of it without awk, grep and find gymnastics, we simply
drop _ALL_ driver_overrides. This can have serious sideeffects!

However, since we only trigger the TC/LUR test pipeline with our own
single test, we don't care at the moment and prefer a simple approach
over some complicated search logic.

One nice sideeffect of this is that we found the original problem with
CX52x0. Yeah, there the driver_override was set in the first round
already, so with our unconditional cleanup, we can now merge CX52x0 back
into the normal control flow.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>