test-connectiviy: wait for iperf3 connection
We first tried to ping and start iperf3 client only after the first
successful ping, but that failed once on a CX53x0[1] and even more
mysteriously on CX52x0[2] it failed **after** one iperf run was already
successful.
So now we try to repeat iperf for up to 20 times and reduce the timeout
to one second.

[1] https://git.beckhoff.dev/beckhoff/test-tcos/-/jobs/2461858
[2] https://git.beckhoff.dev/beckhoff/test-tcos/-/jobs/2461854