netdev: support fifo without pci
When using the fifo without pci we don't write to bar2, but need to
write the upper byte of our physical base address to the highest byte
of the fifo register. When sending we need to include our base physical
address. We share the sending function with pci, as we can just provide
a zero offset for pci.